phase1_cassandra/requirements.txt
# Core Dependencies
javalang==0.13.0
jinja2==3.1.2
pyyaml==6.0.1
click==8.1.7
rich==13.7.0

# LLM Integration (Phase 2)
anthropic>=0.18.0

# Configuration & Validation (Phase 2)
pydantic>=2.0.0

# Evaluation & Metrics (Phase 2)
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
mypy==1.7.1
ruff==0.1.7
black==23.12.0

# Type Stubs
types-PyYAML==6.0.12.12
